Cast hr pointer for proper overload.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@1666 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CD3D9Driver.cpp
@@ -2333,7 +2333,7 @@ bool CD3D9Driver::reset()
 		}
 		else
 		{
-			os::Printer::log("Resetting failed due to unknown reason.", core::stringc(hr).c_str(), ELL_WARNING);
+			os::Printer::log("Resetting failed due to unknown reason.", core::stringc((int)hr).c_str(), ELL_WARNING);
 		}
 		return false;
 	}
